requirements.txt
# AI Assistant — web app (FastAPI + browser UI)
# Install: pip install -r requirements.txt
# Run: python main.py  (starts server and opens browser)

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-multipart>=0.0.9

# Existing AI/LLM and file parsing
requests>=2.31.0
pymupdf>=1.23.0
python-docx>=1.1.0
